Added install property to monkey extension
The install property reinstalls the APK before running the monkey test.
If it's set, a dependency is also created on the assemble task to make
sure the APK is built. This property is needed because the install task
doesn't yet work if there are multiple devices attached, which this
plugin is specifically designed to handle.